requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
apscheduler==3.11.0
httpx==0.28.1
pydantic-settings==2.9.1
pyyaml==6.0.2
redis==5.2.1
